Add a pre-processing step that purges `ghcup`.
The version of `ghc` that comes pre-installed on runner images appears
to be missing documentation for core libraries. This makes it impossible
to build Haddock documentation with working hyperlinks, as `haddock` is
unable to find documentation for those libraries, producing the
following error:

```
Warning: The documentation for the following packages are not installed.
No links will be generated to these packages:
base-4.17.0.0, ghc-bignum-1.3, ghc-prim-0.9.0
```

As a workaround, we create a pre-processing step that calls `ghcup nuke`
to remove this installation. This enables the `haskell/actions/setup@v2`
action to install `ghc` from scratch, together with the documentation
for core libraries.

Of course, this comes at a cost of slightly longer CI times, as we can
no longer take advantage of a pre-installed `ghc`. However, it does make
it possible to build working Haddock documentation.